Rename data to domain
