+++ app/src/main/java/org/simonolander/lambda/MainActivity.kt
@@ -4,14 +4,10 @@ import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
@@ -21,9 +17,8 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import org.simonolander.lambda.data.Level
-import org.simonolander.lambda.data.LevelId
-import org.simonolander.lambda.engine.parse
+import org.simonolander.lambda.domain.Level
+import org.simonolander.lambda.domain.LevelId
 import org.simonolander.lambda.ui.*
 import org.simonolander.lambda.ui.theme.LambdaTheme
