+++ app/src/main/java/org/simonolander/lambda/content/dialog/ApplicationSyntax.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import org.simonolander.lambda.data.DialogBuilder
+import org.simonolander.lambda.domain.DialogBuilder
 import org.simonolander.lambda.misc.javascript
 import org.simonolander.lambda.misc.lambdaCalculus
 import org.simonolander.lambda.ui.theme.LambdaTheme
