+++ app/src/main/java/org/simonolander/lambda/content/exercise/Cardinal.kt
@@ -2,11 +2,10 @@ package org.simonolander.lambda.content.exercise
 
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.withStyle
-import org.simonolander.lambda.data.DialogBuilder
-import org.simonolander.lambda.data.Exercise
-import org.simonolander.lambda.data.TestCase
+import org.simonolander.lambda.domain.DialogBuilder
+import org.simonolander.lambda.domain.Exercise
+import org.simonolander.lambda.domain.TestCase
 import org.simonolander.lambda.engine.*
-import org.simonolander.lambda.misc.lambda
 import org.simonolander.lambda.ui.theme.codeStyle
 
 val cardinalExercise = run {
