+++ app/src/main/java/org/simonolander/lambda/content/exercise/Kite.kt
@@ -2,10 +2,9 @@ package org.simonolander.lambda.content.exercise
 
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.withStyle
-import org.simonolander.lambda.data.DialogBuilder
-import org.simonolander.lambda.data.Exercise
-import org.simonolander.lambda.data.TestCase
-import org.simonolander.lambda.misc.lambdaCalculus
+import org.simonolander.lambda.domain.DialogBuilder
+import org.simonolander.lambda.domain.Exercise
+import org.simonolander.lambda.domain.TestCase
 import org.simonolander.lambda.ui.theme.codeStyle
 
 val kiteExercise = run {
