+++ app/src/main/java/org/simonolander/lambda/domain/Chapter.kt
@@ -1,4 +1,4 @@
-package org.simonolander.lambda.data
+package org.simonolander.lambda.domain
 
 enum class Chapter(
     val id: ChapterId,
