+++ app/src/main/java/org/simonolander/lambda/domain/ChapterId.kt
@@ -1,4 +1,4 @@
-package org.simonolander.lambda.data
+package org.simonolander.lambda.domain
 
 @JvmInline
 value class ChapterId(val value: String)
