+++ app/src/main/java/org/simonolander/lambda/domain/ChatViewModel.kt
@@ -1,4 +1,4 @@
-package org.simonolander.lambda.data
+package org.simonolander.lambda.domain
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
