+++ app/src/main/java/org/simonolander/lambda/domain/LevelId.kt
@@ -1,4 +1,4 @@
-package org.simonolander.lambda.data
+package org.simonolander.lambda.domain
 
 @JvmInline
 value class LevelId(val value: String)
