+++ app/src/main/java/org/simonolander/lambda/ui/levels/LevelView.kt
@@ -10,8 +10,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
-import org.simonolander.lambda.data.Exercise
-import org.simonolander.lambda.data.andExercise
+import org.simonolander.lambda.domain.Exercise
+import org.simonolander.lambda.domain.andExercise
 import org.simonolander.lambda.engine.Expression
 import org.simonolander.lambda.ui.ExecutionState
 import org.simonolander.lambda.ui.ExecutionView
